Added Web and Desktop Support
For web applications, dart: html is used, for mobile, dart: io.
Platform.is is from Dart.io and prevents this library from being used in web projects. To correctly identify if a device is android, iOS, desktop or web, and display the default transitions according to the device used, Get created a selective multiplatform wrapper, at run time and compilation, which is GetPlatform, and the API was left exposed for manipulation of navigation services and personalized dialogs by platform, just like this one.
Swapping Platform.is for GetPlatform.is I can run the sample project both on the Web, on Mobile, and on the desktop.