Added @required for necessary arguments without which an error would be given
I am a beginner in flutter and have started following your architecture (Love it!). But one problem I faced is that when using functions not used in the tutorial I am not sure what all arguments are necessary to pass and what all are optional. I believe that adding @required tag is an easy way of solving this problem for many people. Since I am new to it I may have missed/added the tag in some necessary/unnecessary places.