Change ConfigService to support XDG_CONFIG_HOME
- Replaced isConfigFileAvailable method with resolveConfigFile to look for the
configuration file in different directories depending on the case or priority
- Modified loadConfig to accept a nullable String instead of using a default
value.

These changes allow to have Stacked config file in three different places. Is
possible to have a global configuration for any case on XDG_CONFIG_HOME and
a per project configuration, depending on how stacked command is executed,
with output path of the project or, without it, meaning the project
referenced is on the current directory.